Refactored authentication and user routes: Moved validatePassword and resetPassword to authRoutes; created userRoutes and userController.
