+++ src/app.js
@@ -3,6 +3,7 @@ const connectDB = require("../config/db");
 const medicationRoutes = require("./routes/medicationRoutes.js");
 require("dotenv").config();
 
+const userRoutes = require("./routes/userRoutes");
 const authRoutes = require("./routes/authRoutes");
 const app = express();
 const cors = require("cors");
@@ -10,7 +11,6 @@ const favicon = require("express-favicon");
 const logger = require("morgan");
 
 const mainRouter = require("./routes/mainRouter.js");
-const testRoutes = require("./routes/testRoutes");
 const errorHandler = require("./middleware/errorHandler.js");
 
 // Connect to MongoDB
@@ -26,11 +26,9 @@ app.use(favicon(__dirname + "/public/favicon.ico"));
 
 // routes
 app.use("/api/v1", mainRouter);
-app.use("/api/v1/test", testRoutes);
 app.use("/api/v1/auth", authRoutes);
-app.use("/api/v1", mainRouter);
-app.use("/api/v1/test", testRoutes);
 app.use("/api/v1/inventory", medicationRoutes);
+app.use("/api/v1/users", userRoutes);
 
 app.use(errorHandler);
